[PATCH release] Send reaction out
